chore(tools): Install jp-tools via cargo install
The previous method of invoking the project tools using `cargo run`
created a tight coupling between the tools and the project's source
code. If the code was in a non-compilable state, which can occur during
development, the tools would fail to run, blocking any further automated
actions.

This change decouples the tools by installing them as a pre-compiled
binary using `cargo install`. A new `just install-tools` command has
been added to build and install the `jp-tools` binary from the
`.config/jp/tools` crate.

As a result of this change, developers must now remember to run `just
install-tools` after making any modifications to the tool's source code
to ensure the latest version is used.

Signed-off-by: Jean Mertz <git@jeanmertz.com>