perf(llm, anthropic): Cache Anthropic tool definitions (#229)
This change enables caching for tool definitions in the Anthropic
provider.

The static nature of tool definitions during a conversation, combined
with the sizeable number of tools often used and the number of tokens
tool definitions can take (due to the embedded descriptions and examples
to guide tool use), makes them a prime candidate for caching.

Caching the definitions makes the initial query a bit more expensive,
but subsequent queries are much cheaper.

Signed-off-by: Jean Mertz <git@jeanmertz.com>