chore(just): Handle multi-word arguments in `commit` recipe
The `commit` recipe in the justfile incorrectly handled arguments
containing spaces. The use of `{{ARGS}}` caused the shell to split the
arguments, leading to unexpected behavior when passing a prompt to jp.

This change updates the recipe to use the `positional-arguments`
attribute and the `$@` variable to correctly forward all arguments to
the underlying script. This ensures that quoted multi-word prompts are
treated as a single argument.

We should make this the default behavior for all recipes, but that can
wait for now.

See also: <https://github.com/casey/just/issues/1367>

Signed-off-by: Jean Mertz <git@jeanmertz.com>