feat(cli): Display git hash in `--version` output
Enhance the CLI's version information by embedding the corresponding git
commit hash into the binary at compile time. This provides more precise
version details, which is valuable for debugging and support.

This is achieved by adding a build script to the `jp_cli` crate. The
script gets the short git hash and makes it available through the
`long_version` attribute in clap, which is displayed when using the
`--version` flag. The output of the short `-V` flag remains unchanged.

Signed-off-by: Jean Mertz <git@jeanmertz.com>