refactor(llm, anthropic): Improve message structure and caching
The Anthropic provider now properly uses system content for instructions
and attachments instead of injecting them as user messages with fake
assistant acknowledgments. This change aligns with the best-practices as
described in the Anthropic docs[1].

Additionally, the previous implementation sometimes caused more than
four cache points to be used, which resulted in an API error from
Anthropic. This is now fixed.

[1]: https://docs.anthropic.com/en/docs/build-with-claude/prompt-engineering/overview

Signed-off-by: Jean Mertz <git@jeanmertz.com>