feat(attachment, config): Add description support for attachments (#260)
Attachments can now include an optional descriptions that provide
context about what the attachment represents. The description is
specific per attachment handler, and can be static or dynamic.

For example, the `cmd` handler can include a description of the command
that produced the attachment, by specifying the `description` parameter
in the attachment configuration.

The `bear` handler provides a static description to explain what "bear"
means in the context of the attachment.

---------

Signed-off-by: Jean Mertz <git@jeanmertz.com>