feat(config): Define model IDs as strings
Model IDs can now be defined as strings for convenience. This was
already possible from the command line with the `--model` flag, but
not from the configuration file.

The ID has to have the format `<provider>/<name>`. To make this work
with the support for model aliases, those now no longer accept forward
slashes in the alias name, to distinguish them from the stringified ID
format.

Signed-off-by: Jean Mertz <git@jeanmertz.com>