feat(cli): support piping query data via stdin (#332)
Reads content from `stdin` when piped to `jp`, prepending it to the chat
request. This enables workflows like `cat file.txt | jp q explain`.

When input is piped without a query argument, the editor will open with
the contents of the piped data as the pre-populated query. If a query
argument is provided alongside piped content, the editor is skipped.

---------

Signed-off-by: Jean Mertz <git@jeanmertz.com>