feat(config): overwrite default values during merge (#337)
Add an `is_default` property to `MergedString` and `MergedVec`. Update
the merge logic to completely replace the previous value if it is marked
as default, ignoring any merge strategy defined in the new value.

This prevents user configuration from unintentionally merging with
application defaults, such as the default system prompt or instructions,
ensuring that user settings take precedence.

Mark the default `AssistantConfig` system prompt and instructions as
default values.

---------

Signed-off-by: Jean Mertz <git@jeanmertz.com>