fix(tombmap): prevent dangling pointer in `Mut` wrapper
The `Mut::new` constructor was previously creating a `ptr::NonNull` from
a temporary `HashSet` created on the stack. This resulted in a dangling
pointer that would cause undefined behavior when accessed during
dereferencing or when calling `get_mut`.

The `modified` field in `Mut` is now an `Option`. When created via
`Mut::new`, it is set to `None`, and modification tracking is skipped.
`IterMut` continues to provide the pointer to its internal tracking set
to ensure that modifications via iterator are still correctly captured.

Signed-off-by: Jean Mertz <git@jeanmertz.com>